requirements.txt
roboflow>=1.1.0
ultralytics>=8.0.0
supervision>=0.18.0
opencv-python>=4.8.0
PyYAML>=6.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
python-dotenv>=1.0.0